feat: move types to separate file, add additional css lines
