+++ src/styles/selectMoa.module.css
@@ -27,12 +27,11 @@
 .card {
   box-shadow: 0px 4.4px 4.4px 0px rgba(0, 0, 0, 0.4);
   cursor: pointer;
-  width: 100%;
   max-width: 280px;
   aspect-ratio: 280 / 410;
   border-radius: 20px;
   padding: 20px;
-  transition: 0.3s ease;
+  transition: all 0.3s ease;
 }
 
 .card:hover {
@@ -40,15 +39,30 @@
 }
 /*create-moa 이동 카드*/
 .addCard {
+  background-color: white;
+}
+
+.addBtnConatiner {
+  width: 100%;
+  height: 100%;
   display: flex;
   justify-content: center;
   align-items: center;
-  background-color: white;
 }
 /*moaBox가 담긴 카드*/
 .moaCard {
   background-color: var(--color-gray-200);
 }
+
+/*모아 박스 컨테이너*/
+.moaBoxContainer {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
 /*선택 됐을 때*/
 .selected {
   background-color: var(--color-gray-400);
@@ -81,16 +95,7 @@
   }
 }
 
-/*모아 박스 컨테이너*/
-.moaBoxContainer {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-/*선택하기 버튼 컨테이너너*/
+/*선택하기 버튼 컨테이너*/
 .buttonContainer {
   width: 100%;
   display: flex;
